update: simplify main.py
